requirements.txt
pbr>=0.5.21,<1.0

lockfile>=0.9
python-daemon
python-dateutil
ConfigParser
kombu
python-quantumclient
